+++ recipes/install.rb
@@ -12,6 +12,18 @@
 #(_ P _ ((_ H _ ((_ P _ ((_ - _ ((_ F _ ((_ P _ ((_ M _ (_ 
 #  |_( )__||_( )__||_( )__||_( )__||_( )__||_( )__||_( )__|
 
+#Configure REPO for Debian 6.x
+if node[:platform].include?("debian") && node[:platform_version].include?("6.")
+
+    #Install nginx repo, platform is set inside template
+    cookbook_file "/etc/apt/sources.list.d/dotdeb.list" do
+        source "dotdeb.list"
+        path "/etc/apt/sources.list.d/dotdeb.list"
+        action :create
+    end
+
+end
+
 #Check if we are updating the Repos and System
 if node[:php_fpm][:update_system]
 
@@ -64,7 +76,7 @@ end
 package node[:php_fpm][:package] do
     action :install
 end
- 
+
 #Install PHP Modules if Enabled
 node[:php_fpm][:php_modules].each do |install_packages|
     package install_packages do
@@ -75,6 +87,10 @@ end
 
 #Enable and Restart PHP5-FPM
 service node[:php_fpm][:package] do
+    #Bug in 14.04 for service provider. Adding until resolved.
+    if (platform?('ubuntu') && node['platform_version'].to_f >= 14.04)
+        provider Chef::Provider::Service::Upstart
+    end
     supports :start => true, :stop => true, :restart => true, :reload => true
     action [ :enable, :restart ]
 end