+++ CHANGELOG.md
@@ -25,6 +25,6 @@ php5-fpm CHANGELOG
 
 0.2.2
 -----
-- stajkowski - Updated install receipe to fix the update/upgrade operation.  Now allows for the option and fully functional. Added and tested against more platforms, check .kitchen.yml.
+- stajkowski - Updated install receipe to fix the update/upgrade operation.  Now allows for the option and fully functional. Added and tested against more platforms, check .kitchen.yml.  Fixed 14.04 bug for service provider, will include this until the bug is fixed.
 
 - - -